+++ tests/conftest.py
@@ -1,0 +1,40 @@
+import pytest
+import responses
+from annotationframeworkclient import FrameworkClient
+import os
+from annotationframeworkclient import endpoints
+
+
+TEST_GLOBAL_SERVER = os.environ.get('TEST_SERVER', "https://test.cave.com")
+TEST_LOCAL_SERVER = os.environ.get('TEST_LOCAL_SERVER', "https://local.cave.com")
+TEST_DATASTACK = os.environ.get('TEST_DATASTACK', 'test_stack')
+
+test_info = {
+        "viewer_site": "http://neuromancer-seung-import.appspot.com/",
+        "aligned_volume": {
+            "name": "test_volume",
+            "image_source": f"precomputed://https://{TEST_LOCAL_SERVER}/test-em/v1",
+            "id": 1,
+            "description": "This is a test only dataset."
+        },
+        "synapse_table": "test_synapse_table",
+        "description": "This is the first test datastack. ",
+        "local_server": TEST_LOCAL_SERVER,
+        "segmentation_source": f"graphene://https://{TEST_LOCAL_SERVER}/segmentation/table/test_v1",
+        "soma_table": "test_soma",
+        "analysis_database": None
+    }
+
+
+@pytest.fixture()
+@responses.activate
+def myclient():
+    url_template = endpoints.infoservice_endpoints_v2['datastack_info']
+    mapping = {'i_server_address': TEST_GLOBAL_SERVER,
+               'datastack_name': TEST_DATASTACK}
+    url= url_template.format_map(mapping)
+    responses.add(responses.GET, url,
+                  json=test_info, status=200)
+
+    client = FrameworkClient(TEST_DATASTACK, server_address=TEST_GLOBAL_SERVER)
+    return client
